kutty-bootstrap: Add alert component
Add alert component with support for dismissible and headings.
The API has a read-only dismissible attribute while constructing an Alert.
Note that there are two TODOs that need to be implemented before this can be
called complete:
1. Add JavaScript that makes dismissible work.
2. Add alert-link feature. Links in alert should have this class. Implementing
this is put off for later because all elements might not populate their 'children'
before rendering. This is true especially for some higher order components.