html.HTMLElement: Add support for variadic elements, attrs as dict, and default CLASS
Update the html.HTMLElement to add support during initialization for:
1. Elements as variadic non-keyword arguments
2. Attributes as dict as the first argument
3. A 'CLASS' attribute on subclasses of html.HTMLElement that will be taken as the
default class during initialization.
Note that 'CLASS' itself will always be there in the classes for that element
after initialization (because it is an internal part of that element). Other externally
added classes (either through the dict, or kwargs) will be appended to this.
Another behaviour to note is that during initialization, attributes in kwargs will
take precedence over attributes in dict. This is because kwargs are specified later on
the line and it feels intuitive to have that overwrite previous attributes.