Optional: Allow arbitrary render condition for Optional
Optional now takes an optional argument in its constructor, a
function that tells whether or not the element should be rendered.
The behaviour when that is not given hasn't changed. The default
condition will render when the element is a base element (text or
html), or if the element is a compound element and has children.